streambuf cleanup
git-svn-id: https://svn.code.sf.net/p/sauerbraten/code/src@4979 07af1644-1164-4067-b897-63bf4dc9b748

+++ shared/tools.h
@@ -1117,10 +1117,10 @@ struct streambuf
     streambuf(stream *s) : s(s) {}
     
     T get() { return s->get<T>(); }
-    int get(T *vals, int numvals) { return s->get(vals, numvals); }
+    size_t get(T *vals, size_t numvals) { return s->get(vals, numvals); }
     void put(const T &val) { s->put(&val, 1); }
-    void put(const T *vals, int numvals) { s->put(vals, numvals); } 
-    int length() { return s->size(); }
+    void put(const T *vals, size_t numvals) { s->put(vals, numvals); } 
+    size_t length() { return s->size(); }
 };
 
 enum
